split main interface into 3 others, each for every table
