+++ src/task5/repositories/contactrepository/ContactRepository.java
@@ -1,0 +1,12 @@
+package task5.repositories.contactrepository;
+
+import task5.core.Contact;
+
+import java.util.List;
+
+public interface ContactRepository {
+
+    void register(Contact contact);
+
+    List<Contact> findAllContacts();
+}
